refactor: rename share headers file + change
