+++ bp-daemon/daemon.c
@@ -46,7 +46,7 @@
 #include "netlink.h"
 #include "log.h"
 #include "bp.h"
-#include "../genl_bp.h"
+#include "../common.h"
 
 #define HASHMAP_NUM_BUCKETS 100
 
@@ -357,11 +357,14 @@ void bp_receive_cb(evutil_socket_t fd, short events, void *arg)
 	return;
 }
 
-int send_adu(char *payload, int payload_size, char *eid, int eid_size)
+int bp_send_cb(tls_daemon_ctx_t *ctx, char *payload, int payload_size, char *eid, int eid_size)
 {
 	Sdr sdr;
 	Object bundlePayload;
 	Object bundleZco;
+
+	log_printf(LOG_INFO, "[size=%zu] eid: %s\n", eid_size, eid);
+	log_printf(LOG_INFO, "[size=%zu] payload: %s\n", payload_size, payload);
 
 	sdr = bp_get_sdr();
 	if (sdr == NULL)
@@ -400,146 +403,6 @@ int send_adu(char *payload, int payload_size, char *eid, int eid_size)
 	return 1;
 }
 
-void bundle_cb(tls_daemon_ctx_t *ctx, unsigned long sockid,
-			   char *payload, int payload_size, char *eid, int eid_size)
-{
-	// sock_ctx_t* sock_ctx;
-	// int response = 0; /* Default is success */
-
-	// sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-	// if (sock_ctx == NULL) {
-	// response = -EBADF;
-	// netlink_notify_kernel(ctx, id, response);
-	//	return;
-	//}
-	log_printf(LOG_INFO, "[size=%zu] eid: %s\n", eid_size, eid);
-	log_printf(LOG_INFO, "[size=%zu] payload: %s\n", payload_size, payload);
-	send_adu(payload, payload_size, eid, eid_size);
-
-	// netlink_notify_kernel(ctx, id, response);
-	// return;
-}
-
-// void accept_cb(struct evconnlistener *listener, evutil_socket_t fd,
-// 	struct sockaddr *address, int socklen, void *arg) {
-// 	log_printf(LOG_INFO, "Received connection!\n");
-//
-// 	int port;
-// 	sock_ctx_t* sock_ctx;
-// 	tls_daemon_ctx_t* ctx = arg;
-//
-// 	if (address->sa_family == AF_UNIX) {
-// 		port = strtol(((struct sockaddr_un*)address)->sun_path+1, NULL, 16);
-// 		log_printf(LOG_INFO, "unix port is %05x", port);
-// 	}
-// 	else {
-// 		port = (int)ntohs(((struct sockaddr_in*)address)->sin_port);
-// 	}
-// 	sock_ctx = hashmap_get(ctx->sock_map_port, port);
-// 	if (sock_ctx == NULL) {
-// 		log_printf(LOG_ERROR, "Got an unauthorized connection on port %d\n", port);
-// 		EVUTIL_CLOSESOCKET(fd);
-// 		return;
-// 	}
-// 	log_printf_addr(&sock_ctx->rem_addr);
-//
-// 	if (evutil_make_socket_nonblocking(fd) == -1) {
-// 		log_printf(LOG_ERROR, "Failed in evutil_make_socket_nonblocking: %s\n",
-// 			 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 		EVUTIL_CLOSESOCKET(fd);
-// 		return;
-// 	}
-// 	hashmap_del(ctx->sock_map_port, port);
-// 	//sock_ctx->tls_conn = tls_client_wrapper_setup(sock_ctx->fd, ctx,
-// 	//			sock_ctx->rem_hostname, sock_ctx->is_accepting, sock_ctx->tls_opts);
-//
-// 	return;
-// }
-//
-// void accept_error_cb(struct evconnlistener *listener, void *ctx) {
-//         struct event_base *base = evconnlistener_get_base(listener);
-// #ifndef NO_LOG
-//         int err = EVUTIL_SOCKET_ERROR();
-//         log_printf(LOG_ERROR, "Got an error %d (%s) on the listener\n",
-// 				err, evutil_socket_error_to_string(err));
-// #endif
-//         event_base_loopexit(base, NULL);
-// 	return;
-// }
-//
-// void listener_accept_cb(struct evconnlistener *listener, evutil_socket_t efd,
-// 	struct sockaddr *address, int socklen, void *arg) {
-// 	struct sockaddr_in int_addr = {
-// 		.sin_family = AF_INET,
-// 		.sin_port = 0,
-// 		.sin_addr.s_addr = htonl(INADDR_LOOPBACK)
-// 	};
-// 	socklen_t intaddr_len = sizeof(int_addr);
-// 	sock_ctx_t* sock_ctx = (sock_ctx_t*)arg;
-// 	evutil_socket_t ifd;
-// 	int port;
-// 	sock_ctx_t* new_sock_ctx;
-//         //struct event_base *base = evconnlistener_get_base(listener);
-//
-// 	//log_printf(LOG_DEBUG, "Got a connection on a vicarious listener\n");
-// 	//log_printf_addr(&sock_ctx->int_addr);
-// 	if (evutil_make_socket_nonblocking(efd) == -1) {
-// 		log_printf(LOG_ERROR, "Failed in evutil_make_socket_nonblocking: %s\n",
-// 			 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 		EVUTIL_CLOSESOCKET(efd);
-// 		return;
-// 	}
-//
-// 	new_sock_ctx = (sock_ctx_t*)calloc(1, sizeof(sock_ctx_t));
-// 	if (new_sock_ctx == NULL) {
-// 		return;
-// 	}
-// 	new_sock_ctx->fd = efd;
-// 	//new_sock_ctx->daemon = sock_ctx->daemon;
-// 	//new_sock_ctx->tls_opts = sock_ctx->tls_opts;
-// 	//new_sock_ctx->int_addr = sock_ctx->int_addr;
-// 	//new_sock_ctx->int_addrlen = sock_ctx->int_addrlen;
-//
-// 	ifd = socket(PF_INET, SOCK_STREAM, IPPROTO_TCP);
-// 	if (ifd == -1) {
-// 		return;
-// 	}
-//
-// 	if (bind(ifd, (struct sockaddr*)&int_addr, sizeof(int_addr)) == -1) {
-// 		perror("bind");
-// 		EVUTIL_CLOSESOCKET(ifd);
-// 		return;
-// 	}
-//
-// 	if (getsockname(ifd, (struct sockaddr*)&int_addr, &intaddr_len) == -1) {
-// 		perror("getsockname");
-// 		EVUTIL_CLOSESOCKET(ifd);
-// 		return;
-// 	}
-//
-// 	if (evutil_make_socket_nonblocking(ifd) == -1) {
-// 		log_printf(LOG_ERROR, "Failed in ifd evutil_make_socket_nonblocking: %s\n",
-// 			 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 		EVUTIL_CLOSESOCKET(ifd);
-// 		return;
-// 	}
-//
-// 	port = (int)ntohs((&int_addr)->sin_port);
-// 	hashmap_add(sock_ctx->daemon->sock_map_port, port, (void*)new_sock_ctx);
-// 	return;
-// }
-//
-// void listener_accept_error_cb(struct evconnlistener *listener, void *ctx) {
-//         struct event_base *base = evconnlistener_get_base(listener);
-// #ifndef NO_LOG
-//         int err = EVUTIL_SOCKET_ERROR();
-//         log_printf(LOG_ERROR, "Got an error %d (%s) on a server listener\n",
-// 				err, evutil_socket_error_to_string(err));
-// #endif
-//         event_base_loopexit(base, NULL);
-// 	return;
-// }
-//
 void signal_cb(evutil_socket_t fd, short event, void *arg)
 {
 	int signum = fd; /* why is this fd? */
@@ -557,295 +420,7 @@ void signal_cb(evutil_socket_t fd, short event, void *arg)
 	}
 	return;
 }
-//
-// void socket_cb(tls_daemon_ctx_t* ctx, unsigned long id, char* comm) {
-// 	sock_ctx_t* sock_ctx;
-// 	evutil_socket_t fd;
-// 	int ret;
-// 	int response = 0;
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx != NULL) {
-// 		log_printf(LOG_ERROR, "We have created a socket with this ID already: %lu\n", id);
-// 		netlink_notify_kernel(ctx, id, response);
-// 		return;
-// 	}
-//
-// 	fd = socket(PF_INET, SOCK_STREAM, IPPROTO_TCP);
-// 	if (fd == -1) {
-// 		response = -errno;
-// 	}
-// 	else {
-// 		sock_ctx = (sock_ctx_t*)calloc(1, sizeof(sock_ctx_t));
-// 		if (sock_ctx == NULL) {
-// 			response = -ENOMEM;
-// 		}
-// 		else {
-// 			sock_ctx->id = id;
-// 			sock_ctx->fd = fd;
-// 			//sock_ctx->tls_opts = tls_opts_create(comm);
-// 			hashmap_add(ctx->sock_map, id, (void*)sock_ctx);
-// 		}
-// 	}
-// 	ret = evutil_make_socket_nonblocking(sock_ctx->fd);
-// 	if (ret == -1) {
-// 		log_printf(LOG_ERROR, "Failed in evutil_make_socket_nonblocking: %s\n",
-// 			 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 	}
-//
-// 	log_printf(LOG_INFO, "Socket created on behalf of application %s\n", comm);
-// 	netlink_notify_kernel(ctx, id, response);
-// 	return;
-// }
-//
-// void setsockopt_cb(tls_daemon_ctx_t* ctx, unsigned long id, int level,
-// 		int option, void* value, socklen_t len) {
-// 	sock_ctx_t* sock_ctx;
-// 	int response = 0; /* Default is success */
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		response = -EBADF;
-// 		netlink_notify_kernel(ctx, id, response);
-// 		return;
-// 	}
-//
-// 	// TODO: Call send_peer_auth_req with value: "dest-eid\\text"
-//
-// 	switch (option) {
-// 	default:
-// 		if (setsockopt(sock_ctx->fd, level, option, value, len) == -1) {
-// 			response = -errno;
-// 		}
-// 		break;
-// 	}
-// 	netlink_notify_kernel(ctx, id, response);
-// 	return;
-// }
-//
-// void getsockopt_cb(tls_daemon_ctx_t* ctx, unsigned long id, int level, int option) {
-// 	sock_ctx_t* sock_ctx;
-// 	long value;
-// 	int response = 0;
-// 	char* data = NULL;
-// 	unsigned int len = 0;
-// 	int need_free = 0;
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		netlink_notify_kernel(ctx, id, -EBADF);
-// 		return;
-// 	}
-// 	switch (option) {
-// 	default:
-// 		log_printf(LOG_ERROR, "Default case for getsockopt hit: should never happen\n");
-// 		response = -EBADF;
-// 		break;
-// 	}
-// 	if (response != 0) {
-// 		netlink_notify_kernel(ctx, id, response);
-// 		return;
-// 	}
-// 	netlink_send_and_notify_kernel(ctx, id, data, len);
-// 	if (need_free == 1) {
-// 		free(data);
-// 	}
-// 	return;
-// }
-//
-// void bind_cb(tls_daemon_ctx_t* ctx, unsigned long id, struct sockaddr* int_addr,
-// 	int int_addrlen, struct sockaddr* ext_addr, int ext_addrlen) {
-//
-// 	int ret;
-// 	sock_ctx_t* sock_ctx;
-// 	int response = 0;
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		response = -EBADF;
-// 	}
-// 	else {
-// 		ret = evutil_make_listen_socket_reuseable(sock_ctx->fd);
-// 		if (ret == -1) {
-// 			log_printf(LOG_ERROR, "Failed in evutil_make_listen_socket_reuseable: %s\n",
-// 				 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 			EVUTIL_CLOSESOCKET(sock_ctx->fd);
-// 			return;
-// 		}
-//
-// 		ret = bind(sock_ctx->fd, ext_addr, ext_addrlen);
-// 		if (ret == -1) {
-// 			perror("bind");
-// 			response = -errno;
-// 		}
-// 		else {
-// 			sock_ctx->has_bound = 1;
-// 			sock_ctx->int_addr = *int_addr;
-// 			sock_ctx->int_addrlen = int_addrlen;
-// 			sock_ctx->ext_addr = *ext_addr;
-// 			sock_ctx->ext_addrlen = ext_addrlen;
-// 		}
-// 	}
-// 	netlink_notify_kernel(ctx, id, response);
-// 	return;
-// }
-//
-// void connect_cb(tls_daemon_ctx_t* ctx, unsigned long id, struct sockaddr* int_addr,
-// 	int int_addrlen, struct sockaddr* rem_addr, int rem_addrlen, int blocking) {
-//
-// 	sock_ctx_t* sock_ctx;
-// 	int port;
-//
-// 	if (int_addr->sa_family == AF_UNIX) {
-// 		port = strtol(((struct sockaddr_un*)int_addr)->sun_path+1, NULL, 16);
-// 		log_printf(LOG_INFO, "unix port is %05x", port);
-// 	}
-// 	else {
-// 		port = (int)ntohs(((struct sockaddr_in*)int_addr)->sin_port);
-// 	}
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		netlink_notify_kernel(ctx, id, -EBADF);
-// 		return;
-// 	}
-//
-// 	// do something here
-//
-// 	if (sock_ctx->has_bound == 0) {
-// 		sock_ctx->int_addr = *int_addr;
-// 		sock_ctx->int_addrlen = int_addrlen;
-// 	}
-// 	log_printf(LOG_INFO, "Placing sock_ctx for port %d\n", port);
-// 	hashmap_add(ctx->sock_map_port, port, sock_ctx);
-// 	sock_ctx->rem_addr = *rem_addr;
-// 	sock_ctx->rem_addrlen = rem_addrlen;
-// 	sock_ctx->is_connected = 1; /* is this a lie? */
-//
-// 	if (blocking == 0) {
-// 		log_printf(LOG_INFO, "Nonblocking connect requested\n");
-// 		netlink_notify_kernel(ctx, id, -EINPROGRESS);
-// 	}
-// 	return;
-// }
-//
-// void listen_cb(tls_daemon_ctx_t* ctx, unsigned long id, struct sockaddr* int_addr,
-// 	int int_addrlen, struct sockaddr* ext_addr, int ext_addrlen) {
-//
-// 	int ret;
-// 	sock_ctx_t* sock_ctx;
-// 	int response = 0;
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		response = -EBADF;
-// 	}
-// 	else {
-// 		ret = listen(sock_ctx->fd, SOMAXCONN);
-// 		if (ret == -1) {
-// 			response = -errno;
-// 		}
-// 	}
-// 	netlink_notify_kernel(ctx, id, response);
-// 	if (response != 0) {
-// 		return;
-// 	}
-//
-// 	/* We're done gathering info, let's set up a server */
-// 	ret = evutil_make_socket_nonblocking(sock_ctx->fd);
-// 	if (ret == -1) {
-// 		log_printf(LOG_ERROR, "Failed in evutil_make_socket_nonblocking: %s\n",
-// 			 evutil_socket_error_to_string(EVUTIL_SOCKET_ERROR()));
-// 		EVUTIL_CLOSESOCKET(sock_ctx->fd);
-// 		return;
-// 	}
-//
-// 	//tls_opts_server_setup(sock_ctx->tls_opts);
-// 	sock_ctx->daemon = ctx; /* XXX I don't want this here */
-// 	sock_ctx->listener = evconnlistener_new(ctx->base, listener_accept_cb, sock_ctx,
-// 		LEV_OPT_CLOSE_ON_FREE | LEV_OPT_THREADSAFE, 0, sock_ctx->fd);
-//
-// 	evconnlistener_set_error_cb(sock_ctx->listener, listener_accept_error_cb);
-// 	return;
-// }
-//
-// void associate_cb(tls_daemon_ctx_t* ctx, unsigned long id, struct sockaddr* int_addr, int int_addrlen) {
-// 	sock_ctx_t* sock_ctx;
-// 	int response = 0;
-// 	int port;
-//
-// 	if (int_addr->sa_family == AF_UNIX) {
-// 		port = strtol(((struct sockaddr_un*)int_addr)->sun_path+1, NULL, 16);
-// 		log_printf(LOG_INFO, "unix port is %05x", port);
-// 	}
-// 	else {
-// 		port = (int)ntohs(((struct sockaddr_in*)int_addr)->sin_port);
-// 	}
-// 	sock_ctx = hashmap_get(ctx->sock_map_port, port);
-// 	hashmap_del(ctx->sock_map_port, port);
-// 	if (sock_ctx == NULL) {
-// 		log_printf(LOG_ERROR, "port provided in associate_cb not found");
-// 		response = -EBADF;
-// 		netlink_notify_kernel(ctx, id, response);
-// 		return;
-// 	}
-//
-// 	sock_ctx->id = id;
-// 	sock_ctx->is_connected = 1;
-// 	hashmap_add(ctx->sock_map, id, (void*)sock_ctx);
-//
-// 	//set_netlink_cb_params(sock_ctx->tls_conn, ctx, id);
-// 	//log_printf(LOG_INFO, "Socket %lu accepted\n", id);
-// 	netlink_notify_kernel(ctx, id, response);
-// 	return;
-// }
-//
-// void close_cb(tls_daemon_ctx_t* ctx, unsigned long id) {
-// 	sock_ctx_t* sock_ctx;
-//
-// 	sock_ctx = (sock_ctx_t*)hashmap_get(ctx->sock_map, id);
-// 	if (sock_ctx == NULL) {
-// 		return;
-// 	}
-// 	/* close things here */
-// 	if (sock_ctx->is_accepting == 1) {
-// 		/* This is an ophan server connection.
-// 		 * We don't host its corresponding listen socket
-// 		 * But we were given control of the remote peer
-// 		 * connection */
-// 		hashmap_del(ctx->sock_map, id);
-// 		//tls_opts_free(sock_ctx->tls_opts);
-// 		//free_tls_conn_ctx(sock_ctx->tls_conn);
-// 		free(sock_ctx);
-// 		return;
-// 	}
-// 	if (sock_ctx->is_connected == 1) {
-// 		/* connections under the control of the tls_wrapper code
-// 		 * clean up themselves as a result of the close event
-// 		 * received from one of the endpoints. In this case we
-// 		 * only need to clean up the sock_ctx */
-// 		//netlink_notify_kernel(ctx, id, 0);
-// 		hashmap_del(ctx->sock_map, id);
-// 		//tls_opts_free(sock_ctx->tls_opts);
-// 		//free_tls_conn_ctx(sock_ctx->tls_conn);
-// 		free(sock_ctx);
-// 		return;
-// 	}
-// 	if (sock_ctx->listener != NULL) {
-// 		hashmap_del(ctx->sock_map, id);
-// 		evconnlistener_free(sock_ctx->listener);
-// 		//tls_opts_free(sock_ctx->tls_opts);
-// 		free(sock_ctx);
-// 		//netlink_notify_kernel(ctx, id, 0);
-// 		return;
-// 	}
-// 	hashmap_del(ctx->sock_map, id);
-// 	EVUTIL_CLOSESOCKET(sock_ctx->fd);
-// 	free(sock_ctx);
-// 	//netlink_notify_kernel(ctx, id, 0);
-// 	return;
-// }
-//
+
 /* This function is provided to the hashmap implementation
  * so that it can correctly free all held data */
 void free_sock_ctx(sock_ctx_t *sock_ctx)
@@ -872,53 +447,3 @@ void free_sock_ctx(sock_ctx_t *sock_ctx)
 	free(sock_ctx);
 	return;
 }
-//
-// /* Modified read_fd taken from various online sources. Found without copyright or
-//  * attribution. Examples also in manpages so we could use that if needed */
-// ssize_t recv_fd_from(int fd, void *ptr, size_t nbytes, int *recvfd, struct sockaddr_un* addr, int addr_len) {
-// 	struct msghdr msg;
-// 	struct iovec iov[1];
-// 	ssize_t	n;
-//
-// 	union {
-// 		struct cmsghdr cm;
-// 		char control[CMSG_SPACE(sizeof(int))];
-// 	} control_un;
-// 	struct cmsghdr* cmptr;
-//
-// 	msg.msg_control = control_un.control;
-// 	msg.msg_controllen = sizeof(control_un.control);
-// 	msg.msg_name = addr;
-// 	msg.msg_namelen = addr_len;
-//
-// 	iov[0].iov_base = ptr;
-// 	iov[0].iov_len = nbytes;
-// 	msg.msg_iov = iov;
-// 	msg.msg_iovlen = 1;
-//
-// 	if ((n = recvmsg(fd, &msg, 0)) <= 0) {
-// 		// message length of error or 0
-// 		*recvfd = -1;
-// 		return n;
-// 	}
-//
-// 	if ((cmptr = CMSG_FIRSTHDR(&msg)) != NULL &&
-// 	    cmptr->cmsg_len == CMSG_LEN(sizeof(int))) {
-// 		if (cmptr->cmsg_level != SOL_SOCKET) {
-// 			log_printf(LOG_ERROR, "control level != SOL_SOCKET\n");
-// 			*recvfd = -1;
-// 			return -1;
-// 		}
-// 		if (cmptr->cmsg_type != SCM_RIGHTS) {
-// 			log_printf(LOG_ERROR, "control type != SCM_RIGHTS\n");
-// 			*recvfd = -1;
-// 			return -1;
-// 		}
-// 		*recvfd = *((int *) CMSG_DATA(cmptr));
-// 	}
-// 	else {
-// 		*recvfd = -1; /* descriptor was not passed */
-// 	}
-// 	return n;
-// }
-//
